+++ AlphaBuild/Processing/Renderer.java
@@ -74,11 +74,11 @@ public class Renderer implements Observer {
 
                 gl.glTranslatef(-2f, 2f, -12f);
 
-                float[][][] rgbColor = model.getCubeColors();
+                double[][][] rgbColor = model.getCubeColors();
 
                 for (int i = 0; i < 9; i++) {
                     gl.glBegin(GL2.GL_QUADS);
-                    gl.glColor4f(rgbColor[0][i][0], rgbColor[0][i][1], rgbColor[0][i][2], 1f);
+                    gl.glColor4f((float) rgbColor[0][i][0], (float) rgbColor[0][i][1], (float) rgbColor[0][i][2], 1f);
                     gl.glVertex3f(-1.0f, 1.0f, 0.0f);      // Top left
                     gl.glVertex3f(1.0f, 1.0f, 0.0f);       // Top right
                     gl.glVertex3f(1.0f, -1.0f, 0.0f);      // Bottom right
